Splitting up structs into individual files
