+++ lib/Clause.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <vector>
+
+#include "Literal.h"
+
+struct Clause {
+    std::vector<Literal> literals;
+};
