+++ lib/Function.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "Term.h"
+
+struct Function {
+    std::string name;
+    std::vector<Term> args = {};
+};
