+++ lib/Variable.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "Symbol.h"
+
+struct Variable : Symbol {};
